+++ CHANGELOG.md
@@ -36,6 +36,7 @@ The following changes have been completed, and will be included in the next tagg
   - added FormatMediaTypePartsRuleset
   - added formatMediaTypeParts()
   - added MediaTypePartsToLowerCase ruleset
+  - added MediaTypePartsToUpperCase ruleset
   - added normaliseMediaTypeParts()
   
 ### Fixes